+++ decidim-core/app/helpers/decidim/application_helper.rb
@@ -6,6 +6,7 @@ module Decidim
     include Decidim::OmniauthHelper
     include Decidim::ScopesHelper
     include Decidim::ContextualHelpHelper
+    include Decidim::AmendmentsHelper
 
     # Truncates a given text respecting its HTML tags.
     #