+++ decidim-dev/lib/decidim/dev/test/rspec_support/component.rb
@@ -51,6 +51,7 @@ module Decidim
       include Decidim::DataPortability
       include Searchable
       include Paddable
+      include Amendable
 
       searchable_fields(
         scope_id: { scope: :id },
@@ -58,6 +59,11 @@ module Decidim
         A: [:title],
         D: [:address],
         datetime: :published_at
+      )
+
+      amendable(
+        fields: [:title],
+        form: "Decidim::DummyResources::DummyResourceForm"
       )
 
       component_manifest_name "dummy"
@@ -106,6 +112,7 @@ Decidim.register_component(:dummy) do |component|
     settings.attribute :dummy_global_attribute_1, type: :boolean
     settings.attribute :dummy_global_attribute_2, type: :boolean
     settings.attribute :enable_pads_creation, type: :boolean, default: false
+    settings.attribute :amendments_enabled, type: :boolean, default: false
   end
 
   component.settings(:step) do |settings|
@@ -145,6 +152,7 @@ RSpec.configure do |config|
       unless ActiveRecord::Base.connection.data_source_exists?("decidim_dummy_resources_dummy_resources")
         ActiveRecord::Migration.create_table :decidim_dummy_resources_dummy_resources do |t|
           t.string :title
+          t.string :body
           t.text :address
           t.float :latitude
           t.float :longitude