+++ decidim-proposals/app/models/decidim/proposals/proposal.rb
@@ -21,8 +21,15 @@ module Decidim
       include Decidim::DataPortability
       include Decidim::Hashtaggable
       include Decidim::Proposals::ParticipatoryTextSection
+      include Decidim::Amendable
 
       fingerprint fields: [:title, :body]
+
+      amendable(
+        fields: [:title, :body],
+        ignore: [:published_at, :reference, :state, :answered_at, :answer],
+        form:   "Decidim::Proposals::ProposalForm"
+      )
 
       component_manifest_name "proposals"
 
@@ -195,7 +202,7 @@ module Decidim
       # user - the user to check for authorship
       def editable_by?(user)
         return true if draft?
-        authored_by?(user) && !answered? && within_edit_time_limit? && !copied_from_other_component?
+        !answered? && within_edit_time_limit? && !copied_from_other_component? && created_by?(user)
       end
 
       # Checks whether the user can withdraw the given proposal.